New transfer sources for PID binding tests
Added two new transfer sources under TestTransfers. Both contain the same
hierarchical file structure (including an empty directory) so that the binding
of PIDs to directories can be tested. One contains compressed directories (.7z
files) and the other does not. An empty directory is effected via a
directory with a single .DS_Store file in it (that Archivematica is
expected to delete during Transfer).