Increase the complexity of Demo Transfers
This commit adds a sub-directory to the popular demo transfer
sampledata and updates the DC metadata accordingly.

Notably, the JSON version of the metadata has been updated to
preserve the original repeating dc:subject fields where this
information was lost. JSON requires fields to be repeated using
JSON arrays. If this isn't done it will be silently lost by parsers
such as python's json.tool which is what happened in this instance
here.