Update checksum examples
This commit updates the sample data to be compatible with checksum
verification changes in Archivematica 1.10.

For examples demonstrating multiple checksum algorithms additional
checksums have been added to match the new capability of the system,
for example, SHA512 and BLAKE2 chacksum files have been added.

Where we are demonstrating the failure conditions I have tried to
engineer a failure for each of the checksum types. Failures may
include missing files, and invalid, or incorrect checksums. The goal
is for the microservice job to tell us this!

In the final few changes here, individual checksum files have been
updated so they continue to work with the microservice. The virus
scanning checksum file is a good example.