Add sample for ArchivesSpace matching CSV workflow
This adds a sample transfer for the ArchivesSpace matching CSV workflow,
where a user creates a CSV file called `archivesspaceids.csv` (placed in
the metadata directory) which Archivematica uses to automatically pair
digital objects with resources in ArchivesSpace.

I've included a README which describes how to run the transfer. It also
notes that users may need to change the reference ID in the CSV file to
match a resource in their ArchivesSpace instance.

Connected to archivematica/Issues#258