adds rightsTransfer sample data (#77)
* adds rightsTransfer test data

When updating the Metadata Import documentation, there was a rightsTransfer test being used, but no sample data existed to create the PREMIS rights events. I recreated the transfer and am now adding it to the sample data to make future updates to the docs easier.

* adds rightsTransfer test data

When updating the Metadata Import documentation, there was a
rightsTransfer test being used, but no sample data existed to
create the PREMIS rights events. I recreated the transfer and
am now adding it to the sample data to make future updates to the
docs easier.

* Add Statute row to csv

Also removed citation strings from copyright and license statements

* Fixed spelling in rights.csv

Co-authored-by: sarah-mason <sarah.mason@bodleian.ox.ac.uk>